Yoshiki

+++ ver1/simulator/exec.cc
@@ -28,7 +28,7 @@ void exec_normal_code(unsigned int code, int pc, unsigned int* reg, float* freg,
 			}else{
 				immediate = (int)(code&0b1111111111111111);
 			}
-			reg[rt] = reg[rs] + immediate;
+			reg[rt] = (unsigned int)((int)reg[rs] + immediate);
 			break;
 		case 0b000100 :
     	//BEQ命令の実行
@@ -47,9 +47,9 @@ void exec_normal_code(unsigned int code, int pc, unsigned int* reg, float* freg,
     	rt = (int)((code >> 16) & 0b11111);
     	//*nowの値はそのあとでnow++されるのでここで1を引いとかなければならない
     	if((code>>15)&0b1){
-     		if(reg[rs] > reg[rt]) { *now = *now + (int)(code&0b111111111111111) -power(2,15)- 1; }
+     		if((int)reg[rs] > (int)reg[rt]) { *now = *now + (int)(code&0b111111111111111) -power(2,15)- 1; }
     	}else{
-      	if(reg[rs] > reg[rt]) { *now = *now + (int)(code&0b1111111111111111) - 1; }
+      	if((int)reg[rs] > (int)reg[rt]) { *now = *now + (int)(code&0b1111111111111111) - 1; }
     	}
 			break;
 		case 0b000001 :
@@ -58,9 +58,9 @@ void exec_normal_code(unsigned int code, int pc, unsigned int* reg, float* freg,
     	rt = (int)((code >> 16) & 0b11111);
     	//*nowの値はそのあとでnow++されるのでここで1を引いとかなければならない
     	if((code>>15)&0b1){
-    	  if(reg[rs] < reg[rt]) { *now = *now + (int)(code&0b111111111111111) -power(2,15)- 1; }
+    	  if((int)reg[rs] < (int)reg[rt]) { *now = *now + (int)(code&0b111111111111111) -power(2,15)- 1; }
    	 	}else{
-   	   if(reg[rs] < reg[rt]) { *now = *now + (int)(code&0b1111111111111111) - 1; }
+   	   if((int)reg[rs] < (int)reg[rt]) { *now = *now + (int)(code&0b1111111111111111) - 1; }
    	 	}
 			break;
 		case 0b000101 :
